minor code refactor and add type
